src/Backend-Kafka/Cloud_stream/requirements.txt
eventlet==0.36.1
Flask==2.2.2
Flask_Cors==4.0.1
Flask_SocketIO==5.3.6
kafka_python==2.0.2
pandas==2.2.2
PyMySQL==1.1.1
